+++ src/traci-server/TraCIServer.cpp
@@ -11,7 +11,7 @@
 /// @author  Michael Behrisch
 /// @author  Mario Krumnow
 /// @date    2007/10/24
-/// @version $Id: TraCIServer.cpp 21202 2016-07-19 13:40:35Z behrisch $
+/// @version $Id: TraCIServer.cpp 21813 2016-10-26 20:30:04Z behrisch $
 ///
 /// TraCI server used to control sumo by a remote TraCI client (e.g., ns2)
 /****************************************************************************/
@@ -592,7 +592,7 @@ TraCIServer::initialiseSubscription(const TraCIServer::Subscription& s) {
                         s.contextVars == i->contextVars && s.contextDomain == i->contextDomain && s.range == i->range) {
                     std::vector<std::vector<unsigned char> >::const_iterator k = s.parameters.begin();
                     for (std::vector<int>::const_iterator j = s.variables.begin(); j != s.variables.end(); ++j, ++k) {
-                        const int offset = std::find(i->variables.begin(), i->variables.end(), *j) - i->variables.begin();
+                        const int offset = (int)(std::find(i->variables.begin(), i->variables.end(), *j) - i->variables.begin());
                         if (offset == (int)i->variables.size() || i->parameters[offset] != *k) {
                             i->variables.push_back(*j);
                             i->parameters.push_back(*k);
